Move is_same_graph_with_merge and is_same_graph to aesara.graph.opt_utils
